support auto generate static for uniform (uniform_random)
