+++ paddle/fluid/operators/uniform_random_batch_size_like_op.cu
